JIT: Assume allocations succeed (#82961)
We call a NOMEM function when allocations fail that is already marked as
no-return, but at least MSVC is not able to pick up on this fact. The
result is that all operator new calls end up with an unnecessary null
check. This adds a COMPILER_ASSUME to allow compilers to recognize
that allocations never fail. 

There was also a workaround for an old VSW bug that I've removed.