Remove the dotnet (1) man page (#78229)
All the man pages, including a different version of dotnet.1, are
currently maintained in the dotnet/sdk repo. See
https://github.com/dotnet/sdk/tree/main/documentation/manpages and in
particular, the documentation on how those are generated from
dotnet/docs repository:
https://github.com/dotnet/sdk/blob/main/documentation/manpages/tool/README.md

This leftover copy of dotnet.1 in this repository is extremely out of
date and unmaintained. Lets just delete it.

Co-authored-by: Andy Gocke <angocke@microsoft.com>