Reduce size of async state machine by a reference field (#83696)
Every async state machine today has a field for an Action.  That field is used to cache an Action that's lazily created to point to its MoveNext method.  It's only needed, however, if the state machine awaits something that's not a known awaiter.

Interestingly, Task itself has a field for storing a delegate, which is only used today when the Task is created to invoke a delegate (e.g. Task.Run, ContinueWith, etc.).  I've considered that a liability, but I just realized we can use that same field for this async method cached Action as well, making it relevant to almost all tasks, and avoiding the need to have an extra field on the state machine box.

As the m_action on a task impacts the DebuggerDisplay rendering, I've also added a DebuggerDisplay to the state machine box type.  We can improve this further in the future, and also add a DebuggerTypeProxy later if desired.