Enable regex source gen / compiler vectorization of all sets (#83992)
* Enable regex source gen / compiler vectorization of all sets

When emitting TryFindNextPossibleStartingPosition based on a set that's a fixed-distance from the start of the pattern, today we're able to vectorize a variety of cases.  In .NET 7, this was largely limited to very small sets that contained only 1-5 characters.  Earlier in .NET 8, we expanded this to include ranges (e.g. [0-9]) and more impactfully sets containing any number of only ASCII values (e.g. [A-Za-z0-9]).  That still, however, leaves sets that are non-continguous and contain anything non-ASCII, which is extremely common.  For example, the set `[a-z]` with `RegexOptions.IgnoreCase` actually gets expanded to be `[A-Za-z\u212A]` because the Kelvin sign is considered case-equivalent with `k`, and that means for such a set, we wouldn't employ any vectorized helper and would instead just walk character by character.

With this PR, we'lll now employ some vectorization as part of that remaining class of set as well in the compiler and source generator.  In particular, we do a vectorize a search for the ASCII portion of the set along with anything non-ASCII.  If the found result is ASCII, we're done.  If the found result is non-ASCII, then we proceed to walk character by character as was done previously.

* Address PR feedback