[wasm] New instruction value model for jiterpreter (#84866)
This PR overhauls how the jiterpreter and its heuristic decide whether a trace is profitable to JIT, by assigning each opcode a relative value compared to the interpreter. The goal with this is to account for very high-value opcodes like unrollable memmoves and SIMD, along with accounting for optimizations like constant and data_item inlining.