Reduce regex test suite time (#84865)
On my machine, depending on the order in which the tests end up running, the suite takes anywhere between 50 seconds and 130 seconds.  With these changes, it's fairly consistently around 50 seconds.  The majority of the time is spent in building new Regex instances with the source generator. This change just applies to a few of the test classes with _lots_ of test cases the same batching-in-the-source-generator approach we've taken in other test classes.