Remove S.Sec.Crypto.Algorithms package dependency (#85701)
* Remove S.Sec.Crypto.Algorithms package dependency

Contributes to https://github.com/dotnet/runtime/issues/85641

System.Security.Cryptography.Algorithms/4.3.1 is being referenced in a
few .NET Framework builds. The reference to that package is undesirable
as it brings in the entire netstandard 1.x dependency graph.

The only type used from that package is IncrementalHash which isn't
available as a public API in .NET Framework before 4.7.1. Because of
that, polyfill the netfx code in with minor changes to the source to
make the analyzers happy.