Revert "SetThreadDpiAwareness for host error dialog (#81930)" (#85773)
This reverts commit 6f36be56480dc0a24a9713df5d81875a9ca2c980.

There is an issue with the Windows TaskDialog where expanding/collapsing the detail area will result in incorrect resizing of the window when the DPI awareness of the process and thread are not the same. Instead of explicitly making the thread starting the dialog DPI aware, we just leave it whatever the process is set to. This means that if the developer specified a manifest that made their application DPI aware, the dialog will be non-blurry, but if the developer did not, it will remain blurry on high DPI.