Unreachable filter end (#86492)
Fixes the problem described in https://github.com/dotnet/roslyn/issues/67494 for illink.

`endfilter` instruction must be kept if the filter block is kept, even if the instruction itself is unreachable - per the ECMA spec. Fix the branch removal code to preserve the instruction unconditionally.

Adds a test to validate the new behavior.

I also enabled the test to run on NativeAOT, since it has code to handle this as well. We only validate that the compiler will not crash/warn, the Kept validation is disabled, since we can't really inspect IL of the produced methods in AOT.

Also fixes a bug in the illink test infra where the `By` property on `SkipKeptItemsValidationAttribute` was ignored.